+++ src/main/java/com/github/alexthe666/alexsmobs/entity/EntityLobster.java
@@ -221,7 +221,7 @@ public class EntityLobster extends WaterAnimal implements ISemiAquatic, Bucketab
     }
 
     public int getVariant() {
-        return this.entityData.get(VARIANT).intValue();
+        return this.entityData.get(VARIANT);
     }
 
     public void setVariant(int variant) {
