+++ src/main/java/com/github/alexthe666/alexsmobs/entity/EntitySandShot.java
@@ -79,7 +79,7 @@ public class EntitySandShot extends Entity {
     }
 
     public int getVariant() {
-        return this.entityData.get(VARIANT).intValue();
+        return this.entityData.get(VARIANT);
     }
 
     public void setVariant(int variant) {
